requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.7.3=py37h27cfd23_1
async-timeout=3.0.1=py37_0
atari-py=0.2.6=pypi_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py37_0
brotlipy=0.7.0=py37h27cfd23_1003
bsuite=0.3.4=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2020.12.8=h06a4308_0
cachetools=4.2.0=pyhd3eb1b0_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.4=py37h261ae71_0
chardet=3.0.4=py37h06a4308_1003
click=7.1.2=py_0
cloudpickle=1.6.0=pypi_0
cryptography=2.9.2=py37h1ba5d50_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
descartes=1.1.0=pypi_0
dm-env=1.3=pypi_0
dm-tree=0.1.5=pypi_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
frozendict=1.2=pypi_0
future=0.18.2=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.24.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
graphite2=1.3.14=h23475e2_0
grpcio=1.31.0=py37hf8bcb03_0
gym=0.18.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.10=py_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
jasper=2.0.14=h07fcdf6_1
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=py37h06a4308_0
matplotlib=3.3.3=pypi_0
mizani=0.7.2=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=4.7.6=py37h7b6447c_1
ncurses=6.2=he6710b0_1
networkx=2.5=pypi_0
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=py_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1i=h27cfd23_0
palettable=3.3.0=pypi_0
pandas=1.1.5=pypi_0
patsy=0.5.1=pypi_0
pcre=8.44=he6710b0_0
pillow=7.2.0=pypi_0
pip=20.3.3=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
plotnine=0.7.1=pypi_0
protobuf=3.13.0.1=py37he6710b0_1
psutil=5.7.2=py37h7b6447c_0
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyglet=1.5.0=pypi_0
pyjwt=2.0.0=py37h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyprind=2.11.2=pypi_0
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=pypi_0
pytorch-cpu=1.1.0=py3.7_cpu_0
pytz=2020.4=pypi_0
pywavelets=1.1.1=pypi_0
readline=8.0=h7b6447c_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rlpyt=0.1.2=dev_0
rsa=4.6=py_0
scikit-image=0.18.0=pypi_0
scipy=1.5.4=pypi_0
setuptools=51.0.0=py37h06a4308_2
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.12.1=pypi_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
termcolor=1.1.0=pypi_0
tifffile=2020.12.8=pypi_0
tk=8.6.10=hbc83047_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd3eb1b0_0
werkzeug=1.0.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yarl=1.5.1=py37h7b6447c_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
